refactor of flask files
